+++ src/components/panel-blocks/index.ts
@@ -46,6 +46,8 @@ export { default as PanelFooter } from './PanelFooter';
 export { default as PanelHeader } from './PanelHeader';
 export { default as PanelSkeleton } from './PanelSkeleton';
 export { default as PanelTabs } from './PanelTabs';
+export { default as SettingsCard } from './SettingsCard';
+export type { SaveState } from './SettingsCard';
 export { default as PhoneField } from './PhoneField';
 export { default as PhoneNumbersField } from './PhoneNumbersField';
 export { default as SignalRow } from './SignalRow';